add backward tests, use cast for non-inplace op
