+++ python/paddle/tensor/stat.py
@@ -34,7 +34,7 @@ from paddle.utils.decorator_utils import (
 from ..base.data_feeder import check_type, check_variable_and_dtype
 from ..common_ops_import import Variable
 from ..framework import LayerHelper, convert_np_dtype_to_dtype_, core
-from .manipulation import cast_
+from .manipulation import cast
 from .math import _get_reduce_axis_with_tensor
 
 if TYPE_CHECKING:
@@ -127,7 +127,7 @@ def mean(
         if not isinstance(dtype, (core.VarDesc.VarType, core.DataType)):
             dtype = convert_np_dtype_to_dtype_(dtype)
         if x.dtype != dtype:
-            x = cast_(x, dtype)
+            x = cast(x, dtype)
 
     if in_dynamic_or_pir_mode():
         return _C_ops.mean(x, axis, keepdim, out=out)
